Module haybale::alloc_utils

Utility functions for performing memory allocation. These may be useful in implementing hooks for various functions that perform memory allocation.

Functions

calloc
Allocate a number of bytes given by a times b, where a and b are Operands. The newly-allocated memory will be initialized to all zeroes.

malloc
Allocate a number of bytes given by the Operand.

realloc
Reallocate the given addr to be at least the number of bytes given by the Operand.

zalloc
Allocate a number of bytes given by the Operand. The newly-allocated memory will be initialized to all zeroes.