Function haybale::find_zero_of_func

pub fn find_zero_of_func<'p>(
    funcname: &str, 
    project: &'p Project, 
    config: Config<'p, BtorBackend>
) -> Result<Option<Vec<SolutionValue>>, String>

Given a function, find values of its inputs such that it returns zero. Assumes that the function takes (some number of) integer and/or pointer arguments, and returns an integer. Pointer arguments will be assumed to be never NULL.

project: The Project (set of LLVM modules) in which symbolic execution should take place. In the absence of function hooks (see Config), we will try to enter calls to any functions defined in the Project.

Returns Ok(None) if there are no values of the inputs such that the function returns zero.

Note: find_zero_of_func() may be of some use itself, but also serves as an example of how you can use the other public functions in the crate.