Function haybale::symex_function

pub fn symex_function<'p, B: Backend>(
    funcname: &str, 
    project: &'p Project, 
    config: Config<'p, B>, 
    params: Option<Vec<ParameterVal>>
) -> Result<ExecutionManager<'p, B>>

Begin symbolic execution of the function named funcname, obtaining an ExecutionManager.

project: The Project (set of LLVM modules) in which symbolic execution should take place. In the absence of function hooks (see Config), we will try to enter calls to any functions defined in the Project.

params: a ParameterVal for each parameter to the function, indicating what the initial value of that parameter should be, or if the parameter should be unconstrained (so that the analysis considers all possible values for the parameter). None here is equivalent to supplying a Vec with all ParameterVal::Unconstrained entries.