Function haybale::function_hooks::generic_stub_hook

pub fn generic_stub_hook<B: Backend>(
    _proj: &Project, 
    state: &mut State<B>, 
    call: &dyn IsCall
) -> Result<ReturnValue<B::BV>>

This hook ignores the function arguments and returns an unconstrained value of the appropriate size for the function's return value (or void for void-typed functions).

May be used for functions taking any number and type of arguments, and with any return type.