Function haybale::hook_utils::memset

pub fn memset<B: Backend>(
    state: &mut State<'_, B>, 
    addr: &Operand, 
    val: &Operand, 
    num_bytes: &Operand
) -> Result<B::BV>

Set num_bytes bytes of memory at address addr each to the value val. Each individual byte will be set to val, so only the lowest 8 bits of val will be used.

Returns addr as a BV. Most callers probably won't need this.

Respects the state.config.concretize_memcpy_lengths and state.config.max_memcpy_length settings.