Function haybale::layout::get_offset_bv_index

pub fn get_offset_bv_index<'t, V: BV>(
    base_type: &'t Type, 
    index: &V, 
    solver: V::SolverRef
) -> Result<(V, &'t Type)>

Get the offset (in bytes) of the element at the given index, as well as a reference to the Type of the element at that index.

This function differs from get_offset_constant_index in that it takes an arbitrary BV as index instead of a usize, and likewise returns its offset as a BV.

The result BV will have the same width as the input index.