Function haybale::solver_utils::get_possible_solutions_for_bv

pub fn get_possible_solutions_for_bv<V: BV>(
    solver: V::SolverRef, 
    bv: &V, 
    n: usize
) -> Result<PossibleSolutions<BVSolution>>

Get a description of the possible solutions for the BV.

n: Maximum number of distinct solutions to check for. If there are more than n possible solutions, this returns a PossibleSolutions::AtLeast containing n+1 solutions.

These solutions will be disambiguated - see docs on boolector::BVSolution.

If there are no possible solutions, this returns Ok with an empty PossibleSolutions, rather than returning an Err with Error::Unsat.