Function haybale::solver_utils::max_possible_solution_for_bv_as_u64

pub fn max_possible_solution_for_bv_as_u64<V: BV>(
    solver: V::SolverRef, 
    bv: &V
) -> Result<Option<u64>>

Get the maximum possible solution for the BV: that is, the highest value for which the current set of constraints is still satisfiable. "Maximum" will be interpreted in an unsigned fashion.

Returns Ok(None) if there is no solution for the BV, that is, if the current set of constraints is unsatisfiable. Only returns Err if a solver query itself fails. Panics if the BV is wider than 64 bits.