Struct haybale::State

pub struct State<'p, B: Backend> {
    pub solver: B::SolverRef,
    pub config: Config<'p, B>,
    pub cur_loc: Location<'p>,
    // some fields omitted
}

Fields
solver: B::SolverRef

Reference to the solver instance being used

config: Config<'p, B>

The configuration being used

cur_loc: Location<'p>

Indicates the instruction which is currently being executed

Methods

impl<'p, B: Backend> State<'p, B> where
B: 'p,

pub fn new(
project: &'p Project,
start_loc: Location<'p>,
config: Config<'p, B>
) -> Self

start_loc: the Location where the State should begin executing. As of this writing, start_loc should be the entry point of a function, or you will have problems.

pub fn fork(&self) -> Self

Fully duplicate the State. Unlike with clone(), the State this function returns will have a fully separate (fully duplicated) solver instance. (With clone(), the states will still share references to the same solver instance.)

pub fn sat(&self) -> Result<bool>

Returns true if current constraints are satisfiable, false if not.

Returns Error::SolverError if the query failed (e.g., was interrupted or timed out).

pub fn sat_with_extra_constraints<'b>(
&'b self,
constraints: impl IntoIterator<Item = &'b B::BV>
) -> Result<bool>

Returns true if the current constraints plus the additional constraints conds are together satisfiable, or false if not.

Returns Error::SolverError if the query failed (e.g., was interrupted or timed out).

Does not permanently add the constraints in conds to the solver.

pub fn bvs_must_be_equal(&self, a: &B::BV, b: &B::BV) -> Result<bool>

Returns true if under the current constraints, a and b must have the same value. Returns false if a and b may have different values. (If the current constraints are themselves unsatisfiable, that will result in true.)

A common use case for this function is to test whether some BV must be equal to a given concrete value. You can do this with something like state.bvs_must_be_equal(bv, BV::from_u64(...)).

This function and bvs_can_be_equal() are both more efficient than get_a_solution() or get_possible_solutions()-type functions, as they do not require full model generation. You should prefer this function or bvs_can_be_equal() if they are sufficient for your needs.

pub fn bvs_can_be_equal(&self, a: &B::BV, b: &B::BV) -> Result<bool>

Returns true if under the current constraints, a and b can have the same value. Returns false if a and b cannot have the same value. (If the current constraints are themselves unsatisfiable, that will also result in false.)

A common use case for this function is to test whether some BV can be equal to a given concrete value. You can do this with something like state.bvs_can_be_equal(bv, BV::from_u64(...)).

This function and bvs_must_be_equal() are both more efficient than get_a_solution() or get_possible_solutions()-type functions, as they do not require full model generation. You should prefer this function or bvs_must_be_equal() if they are sufficient for your needs.

pub fn get_a_solution_for_bv(&self, bv: &B::BV) -> Result<Option<BVSolution>>

Get one possible concrete value for the BV. Returns Ok(None) if no possible solution, or Error::SolverError if the solver query failed.

pub fn get_a_solution_for_bv_by_irname(
&mut self,
funcname: &String,
name: &Name
) -> Result<Option<BVSolution>>

Get one possible concrete value for the given IR Name (from the given Function name), which represents a bitvector. Returns Ok(None) if no possible solution, or Error::SolverError if the solver query failed.

pub fn get_possible_solutions_for_bv(
&self,
bv: &B::BV,
n: usize
) -> Result<PossibleSolutions<BVSolution>>

Get a description of the possible solutions for the BV.

n: Maximum number of distinct solutions to check for. If there are more than n possible solutions, this returns a PossibleSolutions::AtLeast containing n+1 solutions.

These solutions will be disambiguated - see docs on boolector::BVSolution.

If there are no possible solutions, this returns Ok with an empty PossibleSolutions, rather than returning an Err with Error::Unsat.

pub fn get_possible_solutions_for_bv_by_irname(
&mut self,
funcname: &String,
name: &Name,
n: usize
) -> Result<PossibleSolutions<BVSolution>>

Get a description of the possible solutions for the given IR Name (from the given Function name), which represents a bitvector.

n: Maximum number of distinct solutions to check for. If there are more than n possible solutions, this returns a PossibleSolutions::AtLeast containing n+1 solutions.

These solutions will be disambiguated - see docs on boolector::BVSolution.

Only returns Err if the solver query itself fails.

pub fn new_bv_with_name(&mut self, name: Name, bits: u32) -> Result<B::BV>

Create a new (unconstrained) BV for the given Name (in the current function).

This function performs uniquing, so if you call it twice with the same Name-Function pair, you will get two different BVs.

Returns the new BV, or Err if it can't be created.

(As of this writing, the only Err that might be returned is Error::LoopBoundExceeded, which is returned if creating the new BV would exceed max_versions_of_name -- see State::new().)

Also, we assume that no two Functions share the same name.

pub fn assign_bv_to_name(&mut self, name: Name, bv: B::BV) -> Result<()>

Assign the given BV to the given Name (in the current function).

This function performs uniquing, so it creates a new version of the variable represented by the (String, Name) pair rather than overwriting the current version.

Returns Err if the assignment can't be performed.

(As of this writing, the only Err that might be returned is Error::LoopBoundExceeded, which is returned if creating the new version of the BV would exceed max_versions_of_name -- see State::new().)

pub fn record_bv_result(
&mut self,
thing: &impl HasResult,
resultval: B::BV
) -> Result<()>

Record the result of thing to be resultval. Assumes thing is in the current function. Will fail with Error::LoopBoundExceeded if that would exceed max_versions_of_name (see State::new).

pub fn overwrite_latest_version_of_bv(&mut self, name: &Name, bv: B::BV)

Overwrite the latest version of the given Name to instead be bv. Assumes Name is in the current function.

pub fn operand_to_bv(&self, op: &Operand) -> Result<B::BV>

Convert an Operand to the appropriate BV. Assumes the Operand is in the current function. (All Operands should be either a constant or a variable we previously added to the state.)

pub fn const_to_bv(&self, c: &Constant) -> Result<B::BV>

Convert a Constant to the appropriate BV.

pub fn get_pointer_to_function(
&self,
funcname: impl Into<String>
) -> Option<&B::BV>

Get a pointer to the given function name. The name will be resolved in the current module.

Returns None if no function was found with that name.

pub fn get_pointer_to_function_hook(&self, funcname: &str) -> Option<&B::BV>

Get a pointer to the currently active hook for the given function name.

Returns None if no function was found with that name, or if there is no currently active hook for that function.

pub fn read(&self, addr: &B::BV, bits: u32) -> Result<B::BV>

Read a value bits bits long from memory at addr. Note that bits can be arbitrarily large.

pub fn write(&mut self, addr: &B::BV, val: B::BV) -> Result<()>

Write a value into memory at addr. Note that val can be an arbitrarily large bitvector.

pub fn allocate(&mut self, bits: impl Into<u64>) -> B::BV

Allocate a value of size bits; return a pointer to the newly allocated object

pub fn get_allocation_size(&mut self, addr: &B::BV) -> Result<Option<u64>>

Get the size, in bits, of the allocation at the given address, or None if that address is not the result of an alloc().

pub fn record_path_entry(&mut self)

Record the current location as a PathEntry in the current path.

pub fn get_path(&self) -> &Vec<PathEntry>

Get the PathEntrys that have been recorded, in order

pub fn push_callsite(&mut self)

Record entering a call at the current location

pub fn pop_callsite(&mut self) -> Option<Location<'p>>

Record leaving the current function. Returns the Location at which the current function was called, or None if the current function was the top-level function.

Also restores the caller's local variables.

pub fn save_backtracking_point(&mut self, bb_to_enter: Name, constraint: B::BV)

Save the current state, about to enter the BasicBlock with the given Name (which must be in the same Module and Function as state.cur_loc), as a backtracking point. The constraint will be added only if we end up backtracking to this point, and only then.

pub fn revert_to_backtracking_point(&mut self) -> Result<bool>

returns Ok(true) if the operation was successful, Ok(false) if there are no saved backtracking points, or Err for other errors

pub fn count_backtracking_points(&self) -> usize

returns the number of saved backtracking points

pub fn pretty_llvm_backtrace(&self) -> String

returns a String containing a formatted view of the current LLVM backtrace

pub fn current_assignments_as_pretty_string(&self) -> Result<String>

returns a String describing a set of satisfying assignments for all variables