Function haybale::symex_function

Important traits for ExecutionManager<'p, B>
impl<'p, B: Backend> Iterator for ExecutionManager<'p, B> where
B: 'p, type Item = Result<ReturnValue<B::BV>>;

pub fn symex_function<'p, B: Backend>(
    funcname: &str, 
    project: &'p Project, 
    config: Config<'p, B>
) -> ExecutionManager<'p, B>

Begin symbolic execution of the function named funcname, obtaining an ExecutionManager. The function's parameters will start completely unconstrained.

project: The Project (set of LLVM modules) in which symbolic execution should take place. In the absence of function hooks (see Config), we will try to enter calls to any functions defined in the Project.