Module haybale::function_hooks

Functions and structures for defining and activating function hooks

Structs

FunctionHooks
A set of function hooks, which will be executed instead of their respective hooked functions if/when the symbolic execution engine encounters a call to one of those hooked functions.

Traits

IsCall
IsCall exists to unify the commonalities between LLVM Call and Invoke instructions

Functions

abort_hook
This hook ignores the function arguments and returns ReturnValue::Abort. It is suitable for hooking functions such as C's exit() which abort the program and never return.

generic_stub_hook
This hook ignores the function arguments and returns an unconstrained value of the appropriate size for the function's return value (or void for void-typed functions).

Type Definitions

Argument
An Argument represents a single argument to a called function, together with zero or more attributes which apply to it