Module haybale::layout

Functions related to the in-memory layout of data types.

Constants

POINTER_SIZE_BITS
our convention is that pointers are 64 bits

Functions

fp_size
Get the size of the FPType, in bits

get_offset_bv_index
Get the offset (in bytes) of the element at the given index, as well as a reference to the Type of the element at that index.

get_offset_constant_index
Get the offset (in bytes) of the element at the given index, as well as the Type of the element at that index.

size
Get the size of the Type, in bits